refactor: extract pure functions from proxy-handler to utils
Move getTransportStatusCode, serializeBlocksForStorage, and
extractLastToolUse to proxy-handler-utils.ts to keep proxy-handler
under the 500-line limit (499 → 441 lines).
